rename proposals tab to governance
