+++ app/src/components/proposals/proposalPage/VotingSidebar/votingCard/VotingCard.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import { CitizenshipQualification } from "@/lib/types"
 import { CardTextProps, VoteType } from "@/components/proposals/proposal.types"
 import { ProposalData } from "@/lib/proposals"
-import { getVotingProps } from "@/app/proposals/utils/votingUtils"
+import { getVotingProps } from "@/app/governance/utils/votingUtils"
 
 const CardText = ({
   proposalData,
